Warning fixes in lnodes, iterate, libsc headers

+++ src/p8est_lnodes.h
@@ -397,7 +397,7 @@ p8est_lnodes_rank_array_index_int (sc_array_t * array, int it)
   P4EST_ASSERT (it >= 0 && (size_t) it < array->elem_count);
 
   return (p8est_lnodes_rank_t *)
-    (array->array + sizeof (p8est_lnodes_rank_t) * it);
+    (array->array + sizeof (p8est_lnodes_rank_t) * (size_t) it);
 }
 
 /** Return a pointer to a lnodes_rank array element indexed by a size_t.
